detekt: fixed checks that the detekt plugin is applied
